+++ qpi-driver/py/qpi_driver/__init__.py
@@ -3,7 +3,7 @@ import importlib.metadata
 try:
     __version__ = importlib.metadata.version("qpi-driver")
 except importlib.metadata.PackageNotFoundError:
-    __version__ = "0.4.2-rc.24"
+    __version__ = "0.4.2-rc.25"
 
 from qpi_driver.builtins import (
     DeviceBuilder,